+++ app/Http/Controllers/Admin/TrashController.php
@@ -80,7 +80,7 @@ class TrashController extends Controller
             $blogPost = BlogPost::onlyTrashed()->findOrFail($id);
 
             //delete post image
-            if(file_exists('assets/uploads/'.$blogPost->image)) {
+            if($blogPost->image && file_exists('assets/uploads/'.$blogPost->image)) {
                 unlink('assets/uploads/'.$blogPost->image);
             }
